requirements.txt
orjson>=3.9
mcp>=0.1.0
pydantic>=2.0
aiogram>=3.0
aiohttp>=3.9
psutil>=5.9
structlog>=23.0
prometheus_client>=0.14
watchdog>=3.0
grpcio>=1.50
grpcio-tools>=1.50
fastapi>=0.100
redis>=4.5
uvicorn>=0.20
duckduckgo_search>=4.0
chromadb>=0.4.0
beautifulsoup4>=4.12
feedparser>=6.0
httpx>=0.24
pyyaml>=6.0

# Phase 4: New integrations
apscheduler>=3.10
discord.py>=2.3
edge-tts>=6.1
docker>=7.0

# GPU acceleration packages require custom build for sm_120 (see install_gpu_deps.ps1 or Dockerfile)
# torch
# cupy
# llama-cpp-python
# vllm
